agrojus/backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
geoalchemy2==0.15.2
psycopg2-binary==2.9.10
alembic==1.14.0

# Geospatial
geopandas==1.0.1
shapely==2.0.6
pyproj==3.7.0
fiona==1.10.1
owslib==0.31.0
rasterio==1.4.3

# HTTP & Scraping
httpx==0.28.1
beautifulsoup4==4.12.3
playwright==1.49.1
lxml==5.3.0

# PDF Generation
reportlab==4.2.5
weasyprint==63.1

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Agro Data Sources
agrobr==1.0.5
basedosdados==2.0.3
google-cloud-bigquery==3.41.0
earthengine-api==1.5.8
inmetpy==0.0.3

# RSS Feed Aggregation
feedparser==6.0.11

# Authentication
PyJWT==2.10.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
